requirements.txt
# Critical packages for the Mirror Agent retrieval graph

# Pydantic for configuration and validation (v2 is required)
pydantic>=2.0.0

# LangChain core abstractions
langchain-core>=0.3.0,<0.4

# Community integrations (provides additional utilities like vectorstores, embeddings, etc.)
langchain-community>=0.3.0,<0.4

# OpenAI integration for query and response models
langchain-openai>=0.2.0,<0.3

# Cohere integration for reranking and embeddings
langchain-cohere>=0.3.0,<0.4
cohere>=2.0.0

# Elasticsearch retriever support (updated version)
langchain-elasticsearch>=0.3.0,<0.4.0

# MongoDB retriever support
langchain-mongodb>=0.2.0,<0.3.0

# Graphing infrastructure for building your retrieval graph
langgraph>=0.0.19

# OpenAI Python package (used by langchain-openai)
openai>=1.0.0

# PyTorch packages for HuggingFace embeddings
torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0

google-generativeai

# Web scraping
playwright>=1.41.0
html5lib>=1.1

# Search engine
duckduckgo-search>=7.2.1

# LLM integration
openai>=1.59.8 # o1 support
anthropic>=0.42.0
python-dotenv>=1.0.0

# Memory management
mem0ai>=0.1.48

# Testing
unittest2>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.5

# gRPC, for Google Generative AI preventing WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
grpcio==1.60.1